Cloudlinux vm addition #1

+++ scripts/provision_post_host.sh
@@ -30,5 +30,6 @@ exe () {
 
 if [[ $1 == "CnC" ]]
 then
-  cp .vagrant/machines/vm-ubuntu/virtualbox/private_key shared/vm-ubuntu.id_rsa
+  [ ! -e .vagrant/machines/vm-ubuntu/virtualbox/private_key ] ||  cp .vagrant/machines/vm-ubuntu/virtualbox/private_key shared/vm-ubuntu.id_rsa
+  [ ! -e .vagrant/machines/vm-clinux/virtualbox/private_key ] ||  cp .vagrant/machines/vm-clinux/virtualbox/private_key shared/vm-clinux.id_rsa
 fi
